Change 'show' split behavior to add custom TopLeftHook() before :topleft
Without it, when the topmost window has a winheight of 0 (in Rolodex mode), Vim somehow makes all window heights equal. I prefer to have the new window open with a minimal height of 1, and keep the other window heights as stable as possible. It's much easier to change the height of the new current window than recreating the previous Rolodex-based layout with the original and the new windows visible.
Move g:previewwindowsplitmode to the front of the command to allow multiple commands joined with cmd1 | winsplitcmd2.