CHG: Stay in the preview window, as the user probably wants to navigate in the opened file
XXX: :pedit uses the CWD of the preview window.
If that already contains a file with another CWD, the shortened command is wrong. Always use the absolute filespec.
FIX: For "tabnr" and "use blank window", must save and use the absolute filespec, as that is not necessarily passed to the :Drop command.